Move format-type-specific bits into populator helpers.
Disklabels are now handled using the same mechanism as all other
formats.
