+++ blivet/populator/helpers/__init__.py
@@ -1,8 +1,13 @@
+from .btrfs import BTRFSFormatPopulator
+from .boot import AppleBootFormatPopulator, EFIFormatPopulator, MacEFIFormatPopulator
 from .disk import DiskDevicePopulator
+from .disklabel import DiskLabelFormatPopulator
 from .dm import DMDevicePopulator
+from .dmraid import DMRaidFormatPopulator
 from .loop import LoopDevicePopulator
-from .lvm import LVMDevicePopulator
-from .mdraid import MDDevicePopulator
+from .luks import LUKSFormatPopulator
+from .lvm import LVMDevicePopulator, LVMFormatPopulator
+from .mdraid import MDDevicePopulator, MDFormatPopulator
 from .multipath import MultipathDevicePopulator
 from .optical import OpticalDevicePopulator
 from .partition import PartitionDevicePopulator
