+++ blivet/populator/populator.py
@@ -32,33 +32,17 @@ gi.require_version("BlockDev", "1.0")
 
 from gi.repository import BlockDev as blockdev
 
-from ..errors import DeviceError, DeviceTreeError, DuplicateVGError, FSError, InvalidDiskLabelError, LUKSError
-from ..devices import BTRFSSubVolumeDevice, BTRFSVolumeDevice, BTRFSSnapShotDevice
+from ..errors import DeviceError, DeviceTreeError
 from ..devices import DMLinearDevice, DMRaidArrayDevice
-from ..devices import FileDevice, LoopDevice, LUKSDevice
-from ..devices import LVMLogicalVolumeDevice, LVMVolumeGroupDevice
-from ..devices import LVMThinPoolDevice, LVMThinLogicalVolumeDevice
-from ..devices import LVMSnapShotDevice, LVMThinSnapShotDevice
+from ..devices import FileDevice, LoopDevice
 from ..devices import MDRaidArrayDevice
-from ..devices import MDContainerDevice
 from ..devices import MultipathDevice
-from ..devices import PartitionDevice
-from ..devices import device_path_to_name
-from ..devices.lvm import get_internal_lv_class
-from .. import formats
 from ..devicelibs import lvm
-from ..devicelibs import raid
 from .. import udev
 from .. import util
 from ..util import open  # pylint: disable=redefined-builtin
 from ..flags import flags
-from ..storage_log import log_exception_info, log_method_call
-from ..i18n import _
-from ..size import Size
-
-from .helpers import DiskDevicePopulator, DMDevicePopulator, LoopDevicePopulator
-from .helpers import LVMDevicePopulator, MDDevicePopulator, MultipathDevicePopulator
-from .helpers import OpticalDevicePopulator, PartitionDevicePopulator
+from ..storage_log import log_method_call
 
 import logging
 log = logging.getLogger("blivet")
@@ -441,582 +425,6 @@ class Populator(object):
             device.original_format = copy.deepcopy(device.format)
         device.device_links = udev.device_get_symlinks(info)
 
-    def handle_udev_disk_label_format(self, info, device):
-        disklabel_type = udev.device_get_disklabel_type(info)
-        log_method_call(self, device=device.name, label_type=disklabel_type)
-        # if there is no disklabel on the device
-        # blkid doesn't understand dasd disklabels, so bypass for dasd
-        if disklabel_type is None and not \
-           (device.is_disk and udev.device_is_dasd(info)):
-            log.debug("device %s does not contain a disklabel", device.name)
-            return
-
-        if device.partitioned:
-            # this device is already set up
-            log.debug("disklabel format on %s already set up", device.name)
-            return
-
-        try:
-            device.setup()
-        except Exception:  # pylint: disable=broad-except
-            log_exception_info(log.warning, "setup of %s failed, aborting disklabel handler", [device.name])
-            return
-
-        # special handling for unsupported partitioned devices
-        if not device.partitionable:
-            try:
-                fmt = formats.get_format("disklabel",
-                                         device=device.path,
-                                         label_type=disklabel_type,
-                                         exists=True)
-            except InvalidDiskLabelError:
-                log.warning("disklabel detected but not usable on %s",
-                            device.name)
-            else:
-                device.format = fmt
-            return
-
-        try:
-            fmt = formats.get_format("disklabel",
-                                     device=device.path,
-                                     exists=True)
-        except InvalidDiskLabelError as e:
-            log.info("no usable disklabel on %s", device.name)
-            if disklabel_type == "gpt":
-                log.debug(e)
-                device.format = formats.get_format(_("Invalid Disk Label"))
-        else:
-            device.format = fmt
-
-    def handle_udev_luks_format(self, info, device):
-        # pylint: disable=unused-argument
-        log_method_call(self, name=device.name, type=device.format.type)
-        if not device.format.uuid:
-            log.info("luks device %s has no uuid", device.path)
-            return
-
-        # look up or create the mapped device
-        if not self.get_device_by_name(device.format.map_name):
-            passphrase = self.__luks_devs.get(device.format.uuid)
-            if device.format.configured:
-                pass
-            elif passphrase:
-                device.format.passphrase = passphrase
-            elif device.format.uuid in self.__luks_devs:
-                log.info("skipping previously-skipped luks device %s",
-                         device.name)
-            elif self._cleanup or flags.testing:
-                # if we're only building the devicetree so that we can
-                # tear down all of the devices we don't need a passphrase
-                if device.format.status:
-                    # this makes device.configured return True
-                    device.format.passphrase = 'yabbadabbadoo'
-            else:
-                # Try each known passphrase. Include luks_devs values in case a
-                # passphrase has been set for a specific device without a full
-                # reset/populate, in which case the new passphrase would not be
-                # in self.__passphrases.
-                for passphrase in self.__passphrases + list(self.__luks_devs.values()):
-                    device.format.passphrase = passphrase
-                    try:
-                        device.format.setup()
-                    except LUKSError:
-                        device.format.passphrase = None
-                    else:
-                        break
-
-            luks_device = LUKSDevice(device.format.map_name,
-                                     parents=[device],
-                                     exists=True)
-            try:
-                luks_device.setup()
-            except (LUKSError, blockdev.CryptoError, DeviceError) as e:
-                log.info("setup of %s failed: %s", device.format.map_name, e)
-                device.remove_child()
-            else:
-                luks_device.update_sysfs_path()
-                self.devicetree._add_device(luks_device)
-                luks_info = udev.get_device(luks_device.sysfs_path)
-                if not luks_info:
-                    log.error("failed to get udev data for %s", luks_device.name)
-                    return
-
-                self.add_udev_device(luks_info, update_orig_fmt=True)
-        else:
-            log.warning("luks device %s already in the tree",
-                        device.format.map_name)
-
-    def handle_vg_lvs(self, vg_device):
-        """ Handle setup of the LV's in the vg_device. """
-        vg_name = vg_device.name
-        lv_info = dict((k, v) for (k, v) in iter(self.devicetree.lv_info.items())
-                       if v.vg_name == vg_name)
-
-        self.names.extend(n for n in lv_info.keys() if n not in self.names)
-
-        if not vg_device.complete:
-            log.warning("Skipping LVs for incomplete VG %s", vg_name)
-            return
-
-        if not lv_info:
-            log.debug("no LVs listed for VG %s", vg_name)
-            return
-
-        all_lvs = []
-        internal_lvs = []
-
-        def add_required_lv(name, msg):
-            """ Add a prerequisite/parent LV.
-
-                The parent is strictly required in order to be able to add
-                some other LV that depends on it. For this reason, failure to
-                add the specified LV results in a DeviceTreeError with the
-                message string specified in the msg parameter.
-
-                :param str name: the full name of the LV (including vgname)
-                :param str msg: message to pass DeviceTreeError ctor on error
-                :returns: None
-                :raises: :class:`~.errors.DeviceTreeError` on failure
-
-            """
-            vol = self.get_device_by_name(name)
-            if vol is None:
-                new_lv = add_lv(lv_info[name])
-                if new_lv:
-                    all_lvs.append(new_lv)
-                vol = self.get_device_by_name(name)
-
-                if vol is None:
-                    log.error("%s: %s", msg, name)
-                    raise DeviceTreeError(msg)
-
-        def add_lv(lv):
-            """ Instantiate and add an LV based on data from the VG. """
-            lv_name = lv.lv_name
-            lv_uuid = lv.uuid
-            lv_attr = lv.attr
-            lv_size = Size(lv.size)
-            lv_type = lv.segtype
-
-            lv_class = LVMLogicalVolumeDevice
-            lv_parents = [vg_device]
-            lv_kwargs = {}
-            name = "%s-%s" % (vg_name, lv_name)
-
-            if self.get_device_by_name(name):
-                # some lvs may have been added on demand below
-                log.debug("already added %s", name)
-                return
-
-            if lv_attr[0] in 'Ss':
-                log.info("found lvm snapshot volume '%s'", name)
-                origin_name = blockdev.lvm.lvorigin(vg_name, lv_name)
-                if not origin_name:
-                    log.error("lvm snapshot '%s-%s' has unknown origin",
-                              vg_name, lv_name)
-                    return
-
-                if origin_name.endswith("_vorigin]"):
-                    lv_kwargs["vorigin"] = True
-                    origin = None
-                else:
-                    origin_device_name = "%s-%s" % (vg_name, origin_name)
-                    add_required_lv(origin_device_name,
-                                    "failed to locate origin lv")
-                    origin = self.get_device_by_name(origin_device_name)
-
-                lv_kwargs["origin"] = origin
-                lv_class = LVMSnapShotDevice
-            elif lv_attr[0] == 'v':
-                # skip vorigins
-                return
-            elif lv_attr[0] in 'IielTCo' and lv_name.endswith(']'):
-                # an internal LV, add the an instance of the appropriate class
-                # to internal_lvs for later processing when non-internal LVs are
-                # processed
-                internal_lvs.append(lv_name)
-                return
-            elif lv_attr[0] == 't':
-                # thin pool
-                lv_class = LVMThinPoolDevice
-            elif lv_attr[0] == 'V':
-                # thin volume
-                pool_name = blockdev.lvm.thlvpoolname(vg_name, lv_name)
-                pool_device_name = "%s-%s" % (vg_name, pool_name)
-                add_required_lv(pool_device_name, "failed to look up thin pool")
-
-                origin_name = blockdev.lvm.lvorigin(vg_name, lv_name)
-                if origin_name:
-                    origin_device_name = "%s-%s" % (vg_name, origin_name)
-                    add_required_lv(origin_device_name, "failed to locate origin lv")
-                    origin = self.get_device_by_name(origin_device_name)
-                    lv_kwargs["origin"] = origin
-                    lv_class = LVMThinSnapShotDevice
-                else:
-                    lv_class = LVMThinLogicalVolumeDevice
-
-                lv_parents = [self.get_device_by_name(pool_device_name)]
-            elif lv_name.endswith(']'):
-                # unrecognized Internal LVM2 device
-                return
-            elif lv_attr[0] not in '-mMrRoOC':
-                # Ignore anything else except for the following:
-                #   - normal lv
-                #   m mirrored
-                #   M mirrored without initial sync
-                #   r raid
-                #   R raid without initial sync
-                #   o origin
-                #   O origin with merging snapshot
-                #   C cached LV
-                return
-
-            lv_dev = self.get_device_by_uuid(lv_uuid)
-            if lv_dev is None:
-                lv_device = lv_class(lv_name, parents=lv_parents,
-                                     uuid=lv_uuid, size=lv_size, seg_type=lv_type,
-                                     exists=True, **lv_kwargs)
-                self.devicetree._add_device(lv_device)
-                if flags.installer_mode:
-                    lv_device.setup()
-
-                if lv_device.status:
-                    lv_device.update_sysfs_path()
-                    lv_device.update_size()
-                    lv_info = udev.get_device(lv_device.sysfs_path)
-                    if not lv_info:
-                        log.error("failed to get udev data for lv %s", lv_device.name)
-                        return lv_device
-
-                    # do format handling now
-                    self.add_udev_device(lv_info, update_orig_fmt=True)
-
-                return lv_device
-
-            return None
-
-        def create_internal_lv(lv):
-            lv_name = lv.lv_name
-            lv_uuid = lv.uuid
-            lv_attr = lv.attr
-            lv_size = Size(lv.size)
-            lv_type = lv.segtype
-
-            matching_cls = get_internal_lv_class(lv_attr)
-            if matching_cls is None:
-                raise DeviceTreeError("No internal LV class supported for type '%s'" % lv_attr[0])
-
-            # strip the "[]"s marking the LV as internal
-            lv_name = lv_name.strip("[]")
-
-            # don't know the parent LV yet, will be set later
-            new_lv = matching_cls(lv_name, vg_device, parent_lv=None, size=lv_size, uuid=lv_uuid, exists=True, seg_type=lv_type)
-            if new_lv.status:
-                new_lv.update_sysfs_path()
-                new_lv.update_size()
-
-                lv_info = udev.get_device(new_lv.sysfs_path)
-                if not lv_info:
-                    log.error("failed to get udev data for lv %s", new_lv.name)
-                    return new_lv
-
-            return new_lv
-
-        # add LVs
-        for lv in lv_info.values():
-            # add the LV to the DeviceTree
-            new_lv = add_lv(lv)
-
-            if new_lv:
-                # save the reference for later use
-                all_lvs.append(new_lv)
-
-        # Instead of doing a topological sort on internal LVs to make sure the
-        # parent LV is always created before its internal LVs (an internal LV
-        # can have internal LVs), we just create all the instances here and
-        # assign their parents later. Those who are not assinged a parent (which
-        # would hold a reference to them) will get eaten by the garbage
-        # collector.
-
-        # create device instances for the internal LVs
-        orphan_lvs = dict()
-        for lv_name in internal_lvs:
-            full_name = "%s-%s" % (vg_name, lv_name)
-            try:
-                new_lv = create_internal_lv(lv_info[full_name])
-            except DeviceTreeError as e:
-                log.warning("Failed to process an internal LV '%s': %s", full_name, e)
-            else:
-                orphan_lvs[full_name] = new_lv
-                all_lvs.append(new_lv)
-
-        # assign parents to internal LVs (and vice versa, see
-        # :class:`~.devices.lvm.LVMInternalLogicalVolumeDevice`)
-        for lv in orphan_lvs.values():
-            parent_lv = lvm.determine_parent_lv(vg_name, lv, all_lvs)
-            if parent_lv:
-                lv.parent_lv = parent_lv
-            else:
-                log.warning("Failed to determine parent LV for an internal LV '%s'", lv.name)
-
-    def handle_udev_lvm_pv_format(self, info, device):
-        # pylint: disable=unused-argument
-        log_method_call(self, name=device.name, type=device.format.type)
-        pv_info = self.devicetree.pv_info.get(device.path, None)
-        if pv_info:
-            vg_name = pv_info.vg_name
-            vg_uuid = pv_info.vg_uuid
-        else:
-            # no info about the PV -> we're done
-            return
-
-        if not vg_name:
-            log.info("lvm pv %s has no vg", device.name)
-            return
-
-        vg_device = self.get_device_by_uuid(vg_uuid, incomplete=True)
-        if vg_device:
-            vg_device.parents.append(device)
-        else:
-            same_name = self.get_device_by_name(vg_name)
-            if isinstance(same_name, LVMVolumeGroupDevice):
-                raise DuplicateVGError("multiple LVM volume groups with the same name (%s)" % vg_name)
-
-            try:
-                vg_size = Size(pv_info.vg_size)
-                vg_free = Size(pv_info.vg_free)
-                pe_size = Size(pv_info.vg_extent_size)
-                pe_count = pv_info.vg_extent_count
-                pe_free = pv_info.vg_free_count
-                pv_count = pv_info.vg_pv_count
-            except (KeyError, ValueError) as e:
-                log.warning("invalid data for %s: %s", device.name, e)
-                return
-
-            vg_device = LVMVolumeGroupDevice(vg_name,
-                                             parents=[device],
-                                             uuid=vg_uuid,
-                                             size=vg_size,
-                                             free=vg_free,
-                                             pe_size=pe_size,
-                                             pe_count=pe_count,
-                                             pe_free=pe_free,
-                                             pv_count=pv_count,
-                                             exists=True)
-            self.devicetree._add_device(vg_device)
-
-        self.handle_vg_lvs(vg_device)
-
-    def handle_udev_md_member_format(self, info, device):
-        # pylint: disable=unused-argument
-        log_method_call(self, name=device.name, type=device.format.type)
-        md_info = blockdev.md.examine(device.path)
-
-        # Use mdadm info if udev info is missing
-        md_uuid = md_info.uuid
-        device.format.md_uuid = device.format.md_uuid or md_uuid
-        md_array = self.get_device_by_uuid(device.format.md_uuid, incomplete=True)
-
-        if md_array:
-            md_array.parents.append(device)
-        else:
-            # create the array with just this one member
-            # level is reported as, eg: "raid1"
-            md_level = md_info.level
-            md_devices = md_info.num_devices
-
-            if md_level is None:
-                log.warning("invalid data for %s: no RAID level", device.name)
-                return
-
-            # md_examine yields metadata (MD_METADATA) only for metadata version > 0.90
-            # if MD_METADATA is missing, assume metadata version is 0.90
-            md_metadata = md_info.metadata or "0.90"
-            md_name = None
-
-            # check the list of devices udev knows about to see if the array
-            # this device belongs to is already active
-            # XXX This is mainly for containers now since their name/device is
-            #     not given by mdadm examine as we run it.
-            for dev in udev.get_devices():
-                if not udev.device_is_md(dev):
-                    continue
-
-                try:
-                    dev_uuid = udev.device_get_md_uuid(dev)
-                    dev_level = udev.device_get_md_level(dev)
-                except KeyError:
-                    continue
-
-                if dev_uuid is None or dev_level is None:
-                    continue
-
-                if dev_uuid == md_uuid and dev_level == md_level:
-                    md_name = udev.device_get_md_name(dev)
-                    break
-
-            md_path = md_info.device or ""
-            if md_path and not md_name:
-                md_name = device_path_to_name(md_path)
-                if re.match(r'md\d+$', md_name):
-                    # md0 -> 0
-                    md_name = md_name[2:]
-
-                if md_name:
-                    array = self.get_device_by_name(md_name, incomplete=True)
-                    if array and array.uuid != md_uuid:
-                        log.error("found multiple devices with the name %s", md_name)
-
-            if md_name:
-                log.info("using name %s for md array containing member %s",
-                         md_name, device.name)
-            else:
-                log.error("failed to determine name for the md array %s", (md_uuid or "unknown"))
-                return
-
-            array_type = MDRaidArrayDevice
-            try:
-                if raid.get_raid_level(md_level) is raid.Container and \
-                   getattr(device.format, "biosraid", False):
-                    array_type = MDContainerDevice
-            except raid.RaidError as e:
-                log.error("failed to create md array: %s", e)
-                return
-
-            try:
-                md_array = array_type(
-                    md_name,
-                    level=md_level,
-                    member_devices=md_devices,
-                    uuid=md_uuid,
-                    metadata_version=md_metadata,
-                    exists=True
-                )
-            except (ValueError, DeviceError) as e:
-                log.error("failed to create md array: %s", e)
-                return
-
-            md_array.update_sysfs_path()
-            md_array.parents.append(device)
-            self.devicetree._add_device(md_array)
-            if md_array.status:
-                array_info = udev.get_device(md_array.sysfs_path)
-                if not array_info:
-                    log.error("failed to get udev data for %s", md_array.name)
-                    return
-
-                self.add_udev_device(array_info, update_orig_fmt=True)
-
-    def handle_udev_dmraid_member_format(self, info, device):
-        # if dmraid usage is disabled skip any dmraid set activation
-        if not flags.dmraid:
-            return
-
-        log_method_call(self, name=device.name, type=device.format.type)
-        name = udev.device_get_name(info)
-        uuid = udev.device_get_uuid(info)
-        major = udev.device_get_major(info)
-        minor = udev.device_get_minor(info)
-
-        # Have we already created the DMRaidArrayDevice?
-        rs_names = blockdev.dm.get_member_raid_sets(uuid, name, major, minor)
-        if len(rs_names) == 0:
-            log.warning("dmraid member %s does not appear to belong to any "
-                        "array", device.name)
-            return
-
-        for rs_name in rs_names:
-            dm_array = self.get_device_by_name(rs_name, incomplete=True)
-            if dm_array is not None:
-                # We add the new device.
-                dm_array.parents.append(device)
-            else:
-                # Activate the Raid set.
-                blockdev.dm.activate_raid_set(rs_name)
-                dm_array = DMRaidArrayDevice(rs_name,
-                                             parents=[device])
-
-                self.devicetree._add_device(dm_array)
-
-                # Wait for udev to scan the just created nodes, to avoid a race
-                # with the udev.get_device() call below.
-                udev.settle()
-
-                # Get the DMRaidArrayDevice a DiskLabel format *now*, in case
-                # its partitions get scanned before it does.
-                dm_array.update_sysfs_path()
-                dm_array_info = udev.get_device(dm_array.sysfs_path)
-                self.handle_udev_disk_label_format(dm_array_info, dm_array)
-
-                # Use the rs's object on the device.
-                # pyblock can return the memebers of a set and the
-                # device has the attribute to hold it.  But ATM we
-                # are not really using it. Commenting this out until
-                # we really need it.
-                # device.format.raidmem = block.getMemFromRaidSet(dm_array,
-                #        major=major, minor=minor, uuid=uuid, name=name)
-
-    def handle_btrfs_format(self, info, device):
-        log_method_call(self, name=device.name)
-        uuid = udev.device_get_uuid(info)
-
-        btrfs_dev = None
-        for d in self.devicetree.devices:
-            if isinstance(d, BTRFSVolumeDevice) and d.uuid == uuid:
-                btrfs_dev = d
-                break
-
-        if btrfs_dev:
-            log.info("found btrfs volume %s", btrfs_dev.name)
-            btrfs_dev.parents.append(device)
-        else:
-            label = udev.device_get_label(info)
-            log.info("creating btrfs volume btrfs.%s", label)
-            btrfs_dev = BTRFSVolumeDevice(label, parents=[device], uuid=uuid,
-                                          exists=True)
-            self.devicetree._add_device(btrfs_dev)
-
-        if not btrfs_dev.subvolumes:
-            snapshots = btrfs_dev.list_subvolumes(snapshots_only=True)
-            snapshot_ids = [s.id for s in snapshots]
-            for subvol_dict in btrfs_dev.list_subvolumes():
-                vol_id = subvol_dict.id
-                vol_path = subvol_dict.path
-                parent_id = subvol_dict.parent_id
-                if vol_path in [sv.name for sv in btrfs_dev.subvolumes]:
-                    continue
-
-                # look up the parent subvol
-                parent = None
-                subvols = [btrfs_dev] + btrfs_dev.subvolumes
-                for sv in subvols:
-                    if sv.vol_id == parent_id:
-                        parent = sv
-                        break
-
-                if parent is None:
-                    log.error("failed to find parent (%d) for subvol %s",
-                              parent_id, vol_path)
-                    raise DeviceTreeError("could not find parent for subvol")
-
-                fmt = formats.get_format("btrfs",
-                                         device=btrfs_dev.path,
-                                         exists=True,
-                                         vol_uuid=btrfs_dev.format.vol_uuid,
-                                         subvolspec=vol_path,
-                                         mountopts="subvol=%s" % vol_path)
-                if vol_id in snapshot_ids:
-                    device_class = BTRFSSnapShotDevice
-                else:
-                    device_class = BTRFSSubVolumeDevice
-
-                subvol = device_class(vol_path,
-                                      vol_id=vol_id,
-                                      fmt=fmt,
-                                      parents=[parent],
-                                      exists=True)
-                self.devicetree._add_device(subvol)
-
     def handle_udev_device_format(self, info, device):
         log_method_call(self, name=getattr(device, "name", None))
 
@@ -1028,131 +436,14 @@ class Populator(object):
             return
 
         name = udev.device_get_name(info)
-        uuid = udev.device_get_uuid(info)
-        label = udev.device_get_label(info)
-        format_type = udev.device_get_format(info)
-        serial = udev.device_get_serial(info)
-
-        is_multipath_member = (device.is_disk and
-                               blockdev.mpath_is_mpath_member(device.path))
-
-        # Now, if the device is a disk, see if there is a usable disklabel.
-        # If not, see if the user would like to create one.
-        # XXX ignore disklabels on multipath or biosraid member disks
-        if not udev.device_is_biosraid_member(info) and \
-           not is_multipath_member and \
-           format_type != "iso9660":
-            self.handle_udev_disk_label_format(info, device)
-            if device.partitioned or self.is_ignored(info) or \
-               (not device.partitionable and
-                    device.format.type == "disklabel"):
-                # If the device has a disklabel, or the user chose not to
-                # create one, we are finished with this device. Otherwise
-                # it must have some non-disklabel formatting, in which case
-                # we fall through to handle that.
-                return
-
-        if (not device) or (not format_type) or device.format.type:
+        if (not device or
+            (not udev.device_get_format(info) and not udev.device_get_disklabel_type(info)) or
+           device.format.type):
             # this device has no formatting or it has already been set up
-            # FIXME: this probably needs something special for disklabels
             log.debug("no type or existing type for %s, bailing", name)
             return
 
-        # set up the common arguments for the format constructor
-        format_designator = format_type
-        kwargs = {"uuid": uuid,
-                  "label": label,
-                  "device": device.path,
-                  "serial": serial,
-                  "exists": True}
-
-        # set up type-specific arguments for the format constructor
-        if format_type == "crypto_LUKS":
-            # luks/dmcrypt
-            kwargs["name"] = "luks-%s" % uuid
-        elif format_type in formats.mdraid.MDRaidMember._udev_types:
-            # mdraid
-            try:
-                # ID_FS_UUID contains the array UUID
-                kwargs["md_uuid"] = udev.device_get_uuid(info)
-            except KeyError:
-                log.warning("mdraid member %s has no md uuid", name)
-
-            # reset the uuid to the member-specific value
-            # this will be None for members of v0 metadata arrays
-            kwargs["uuid"] = udev.device_get_md_device_uuid(info)
-
-            kwargs["biosraid"] = udev.device_is_biosraid_member(info)
-        elif format_type == "LVM2_member":
-            # lvm
-            pv_info = self.devicetree.pv_info.get(device.path, None)
-            if pv_info:
-                if pv_info.vg_name:
-                    kwargs["vg_name"] = pv_info.vg_name
-                else:
-                    log.warning("PV %s has no vg_name", name)
-                if pv_info.vg_uuid:
-                    kwargs["vg_uuid"] = pv_info.vg_uuid
-                else:
-                    log.warning("PV %s has no vg_uuid", name)
-                if pv_info.pe_start:
-                    kwargs["pe_start"] = Size(pv_info.pe_start)
-                else:
-                    log.warning("PV %s has no pe_start", name)
-        elif format_type == "vfat":
-            # efi magic
-            if isinstance(device, PartitionDevice) and device.bootable:
-                efi = formats.get_format("efi")
-                if efi.min_size <= device.size <= efi.max_size:
-                    format_designator = "efi"
-        elif format_type == "hfsplus":
-            if isinstance(device, PartitionDevice):
-                macefi = formats.get_format("macefi")
-                if macefi.min_size <= device.size <= macefi.max_size and \
-                   device.parted_partition.name == macefi.name:
-                    format_designator = "macefi"
-        elif format_type == "hfs":
-            # apple bootstrap magic
-            if isinstance(device, PartitionDevice) and device.bootable:
-                apple = formats.get_format("appleboot")
-                if apple.min_size <= device.size <= apple.max_size:
-                    format_designator = "appleboot"
-        elif format_type == "btrfs":
-            # the format's uuid attr will contain the UUID_SUB, while the
-            # overarching volume UUID will be stored as vol_uuid
-            kwargs["uuid"] = info["ID_FS_UUID_SUB"]
-            kwargs["vol_uuid"] = uuid
-        elif format_type == "multipath_member":
-            # blkid does not care that the UUID it sees on a multipath member is
-            # for the multipath set's (and not the member's) formatting, so we
-            # have to discard it.
-            kwargs.pop("uuid")
-            kwargs.pop("label")
-
-        try:
-            log.info("type detected on '%s' is '%s'", name, format_designator)
-            device.format = formats.get_format(format_designator, **kwargs)
-            if device.format.type:
-                log.info("got format: %s", device.format)
-        except FSError:
-            log.warning("type '%s' on '%s' invalid, assuming no format",
-                        format_designator, name)
-            device.format = formats.DeviceFormat()
-            return
-
-        #
-        # now do any special handling required for the device's format
-        #
-        if device.format.type == "luks":
-            self.handle_udev_luks_format(info, device)
-        elif device.format.type == "mdmember":
-            self.handle_udev_md_member_format(info, device)
-        elif device.format.type == "dmraidmember":
-            self.handle_udev_dmraid_member_format(info, device)
-        elif device.format.type == "lvmpv":
-            self.handle_udev_lvm_pv_format(info, device)
-        elif device.format.type == "btrfs":
-            self.handle_btrfs_format(info, device)
+        # TODO: set up the formatting
 
     def update_device_format(self, device):
         log.info("updating format of device: %s", device)
